+++ rewrite/changelog.md
@@ -19,6 +19,8 @@ The subsections are formatted as they need when publishing a release to the Play
 
 ## Unreleased
 
+ * Added: support for changing network
+
 ## [0.9.31] - 2019-08-30
 
  * Fixed: Binding to service after resuming activity